Add UserUpcomingMatchQueryService and mapper
Introduce UserUpcomingMatchQueryService to handle user-specific upcoming match logic (group/tournament filtering) and keep UpcomingMatchQueryService focused on generic matches. Add UpcomingMatchRestMapper to centralize mapping from UpcomingMatchResult to UpcomingMatchDto and remove the controller's inline toDto. Update MatchController to use the new mapper and user service. Adjust tests: simplify UpcomingMatchQueryServiceTest and add UserUpcomingMatchQueryServiceTest covering group-related behavior.
