+++ backend/src/main/java/com/predictorama/backend/adapter/rest/controller/MatchController.java
@@ -1,10 +1,10 @@
 package com.predictorama.backend.adapter.rest.controller;
 
 import com.predictorama.backend.adapter.rest.SessionService;
-import com.predictorama.backend.adapter.rest.dto.GroupReferenceDto;
 import com.predictorama.backend.adapter.rest.dto.UpcomingMatchDto;
+import com.predictorama.backend.adapter.rest.mapper.UpcomingMatchRestMapper;
 import com.predictorama.backend.domain.service.UpcomingMatchQueryService;
-import com.predictorama.backend.domain.service.result.UpcomingMatchResult;
+import com.predictorama.backend.domain.service.UserUpcomingMatchQueryService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -23,13 +23,14 @@ public class MatchController {
     private static final Logger log = LoggerFactory.getLogger(MatchController.class);
 
     private final UpcomingMatchQueryService upcomingMatchQueryService;
+    private final UserUpcomingMatchQueryService userUpcomingMatchQueryService;
     private final SessionService sessionService;
 
     @GetMapping("/upcoming")
     public List<UpcomingMatchDto> getUpcomingMatches() {
         log.info("GET /api/matches/upcoming");
         return upcomingMatchQueryService.getGenericUpcomingMatches().stream()
-                .map(this::toDto)
+                .map(UpcomingMatchRestMapper::toDto)
                 .toList();
     }
 
@@ -37,29 +38,8 @@ public class MatchController {
     public List<UpcomingMatchDto> getMyUpcomingMatches(HttpSession session) {
         var userId = sessionService.getUserIdOrThrow(session);
         log.info("GET /api/matches/upcoming/my - userId={}", userId);
-        return upcomingMatchQueryService.getUpcomingMatches(userId).stream()
-                .map(this::toDto)
+        return userUpcomingMatchQueryService.getUpcomingMatches(userId).stream()
+                .map(UpcomingMatchRestMapper::toDto)
                 .toList();
-    }
-
-    private UpcomingMatchDto toDto(UpcomingMatchResult result) {
-        var match = result.getMatch();
-        var groups = result.getUserGroups().stream()
-                .map(group -> GroupReferenceDto.builder()
-                        .groupId(group.getId().toString())
-                        .groupName(group.getName())
-                        .competitionId(result.getCompetitionCode())
-                        .build())
-                .toList();
-
-        return UpcomingMatchDto.builder()
-                .matchId(match.getId())
-                .homeTeamName(match.getHomeTeam().getName())
-                .homeTeamImage(match.getHomeTeam().getImageUrl())
-                .awayTeamName(match.getAwayTeam().getName())
-                .awayTeamImage(match.getAwayTeam().getImageUrl())
-                .kickoffTime(match.getKickoffTime())
-                .groups(groups)
-                .build();
     }
 }
